// src/app/core/models/models.ts
export interface Article {
  _id: string;
  Reference: string;
  Designation: string;
  Categorie: string | Category;
  Photo: string;
  description: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}
// ...
export interface Category {
  _id: string;
  description: string; // Include the 'description' property here if it exists
  Libelle: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}